+++ fnpdjango/actions.py
@@ -1,16 +1,7 @@
 # Source: https://gist.github.com/jeremyjbowers/e8d007446155c12033e6
-from __future__ import unicode_literals
-
 import csv
 from django.http import HttpResponse
 from django.utils.translation import ugettext_lazy as _
-
-try:
-    unicode
-except NameError:
-    pass
-else:
-    str = unicode
 
 
 def export_as_csv_action(description=_("Export selected objects as CSV file"), fields=None, exclude=None, header=True):