+++ fnpdjango/management/commands/makecontribmessages.py
@@ -1,9 +1,6 @@
-# -*- coding: utf-8 -*-
 # This file is part of FNPDjango, licensed under GNU Affero GPLv3 or later.
 # Copyright © Fundacja Nowoczesna Polska. See NOTICE for more information.
 #
-from __future__ import print_function
-
 import os
 from optparse import make_option
 from django.core.management.base import BaseCommand