+++ fnpdjango/utils/settings.py
@@ -1,19 +1,9 @@
 """
 Utilities for global settings.
 """
-from django.utils.encoding import python_2_unicode_compatible
 from django.utils.functional import Promise
 
-# Use Python3 str.
-try:
-    unicode
-except NameError:
-    pass
-else:
-    str = unicode
 
-
-@python_2_unicode_compatible
 class LazyUGettextLazy(Promise):
     """You can use it to internationalize strings in settings.