+++ tests/tests/test_utils_settings.py
@@ -1,18 +1,8 @@
-# -*- coding: utf-8 -*-
 # This file is part of FNPDjango, licensed under GNU Affero GPLv3 or later.
 # Copyright © Fundacja Nowoczesna Polska. See README.md for more information.
 #
-from __future__ import unicode_literals
-
 from django.conf import settings
 from django.test import TestCase
-
-try:
-    unicode
-except NameError:
-    pass
-else:
-    str = unicode
 
 
 class UtilsSettingsTestCase(TestCase):