+++ tests/views.py
@@ -1,9 +1,4 @@
 from django.http import HttpResponse
-from django.utils.translation import get_language
-
-
-def get_lang(request):
-    return HttpResponse(get_language())
 
 
 def ip(request):